+++ src/components/Controllers/Metadata/store.ts
@@ -5,45 +5,22 @@ import { assert } from 'ts-essentials'
 import { Client } from '../../../client'
 import { IFile } from '../../../interfaces'
 import { MetadataProps } from './Metadata'
-import { IPublish } from '../../../interfaces/publish'
 
 export interface State {
-  // Data
-
   client: Client
   file: IFile
-  type?: 'package' | 'resource' | 'dialect' | 'schema' | 'checklist' | 'pipeline'
-  onPathChange?: (path?: string) => void
   descriptor?: object
-  isPublish?: boolean
   publishedPath?: string
-
-  // Logic
   loadDescriptor: () => Promise<void>
-  togglePublish: () => void
-  publishPackage: (params: IPublish) => Promise<any>
 }
 
 export function makeStore(props: MetadataProps) {
   return createStore<State>((set, get) => ({
-    // Data
-
     ...props,
-
-    // Logic
-
     loadDescriptor: async () => {
       const { client, file } = get()
       const { data } = await client.dataRead({ path: file.path })
       set({ descriptor: data })
-    },
-    togglePublish: () => {
-      set({ isPublish: !get().isPublish })
-    },
-    publishPackage: async (params: IPublish) => {
-      const { client } = get()
-      const { content } = await client.projectPublish({ params })
-      return content
     },
   }))
 }
